+++ apps/explorer/test/explorer/etherscan_test.exs
@@ -858,7 +858,12 @@ defmodule Explorer.EtherscanTest do
         |> insert()
         |> with_block()
 
-      token_transfer = insert(:token_transfer, transaction: transaction)
+      token_transfer =
+        insert(:token_transfer,
+          transaction: transaction,
+          block: transaction.block,
+          block_number: transaction.block_number
+        )
 
       [found_token_transfer] = Etherscan.list_token_transfers(token_transfer.from_address_hash, nil)
 
@@ -871,7 +876,12 @@ defmodule Explorer.EtherscanTest do
         |> insert()
         |> with_block()
 
-      token_transfer = insert(:token_transfer, transaction: transaction)
+      token_transfer =
+        insert(:token_transfer,
+          transaction: transaction,
+          block: transaction.block,
+          block_number: transaction.block_number
+        )
 
       [found_token_transfer] = Etherscan.list_token_transfers(token_transfer.to_address_hash, nil)
 
@@ -893,9 +903,26 @@ defmodule Explorer.EtherscanTest do
         |> insert()
         |> with_block()
 
-      insert(:token_transfer, from_address: address1, transaction: transaction)
-      insert(:token_transfer, from_address: address1, transaction: transaction)
-      insert(:token_transfer, from_address: address2, transaction: transaction)
+      insert(:token_transfer,
+        from_address: address1,
+        transaction: transaction,
+        block: transaction.block,
+        block_number: transaction.block_number
+      )
+
+      insert(:token_transfer,
+        from_address: address1,
+        transaction: transaction,
+        block: transaction.block,
+        block_number: transaction.block_number
+      )
+
+      insert(:token_transfer,
+        from_address: address2,
+        transaction: transaction,
+        block: transaction.block,
+        block_number: transaction.block_number
+      )
 
       found_token_transfers = Etherscan.list_token_transfers(address1.hash, nil)
 
@@ -914,7 +941,12 @@ defmodule Explorer.EtherscanTest do
         |> insert()
         |> with_block()
 
-      token_transfer = insert(:token_transfer, transaction: transaction)
+      token_transfer =
+        insert(:token_transfer,
+          transaction: transaction,
+          block: transaction.block,
+          block_number: transaction.block_number
+        )
 
       insert(:block)
 
@@ -933,7 +965,12 @@ defmodule Explorer.EtherscanTest do
         |> insert()
         |> with_block()
 
-      token_transfer = insert(:token_transfer, transaction: transaction)
+      token_transfer =
+        insert(:token_transfer,
+          transaction: transaction,
+          block: transaction.block,
+          block_number: transaction.block_number
+        )
 
       {:ok, token} = Chain.token_from_address_hash(token_transfer.token_contract_address_hash)
 
@@ -1049,11 +1086,29 @@ defmodule Explorer.EtherscanTest do
         |> insert()
         |> with_block(third_block)
 
-      second_block_token_transfers = insert_list(2, :token_transfer, from_address: address, transaction: transaction2)
+      second_block_token_transfers =
+        insert_list(2, :token_transfer,
+          from_address: address,
+          transaction: transaction2,
+          block: transaction2.block,
+          block_number: transaction2.block_number
+        )
 
-      first_block_token_transfers = insert_list(2, :token_transfer, from_address: address, transaction: transaction3)
+      first_block_token_transfers =
+        insert_list(2, :token_transfer,
+          from_address: address,
+          transaction: transaction3,
+          block: transaction3.block,
+          block_number: transaction3.block_number
+        )
 
-      third_block_token_transfers = insert_list(2, :token_transfer, from_address: address, transaction: transaction1)
+      third_block_token_transfers =
+        insert_list(2, :token_transfer,
+          from_address: address,
+          transaction: transaction1,
+          block: transaction1.block,
+          block_number: transaction1.block_number
+        )
 
       options1 = %{page_number: 1, page_size: 2}
 
@@ -1106,7 +1161,12 @@ defmodule Explorer.EtherscanTest do
           |> insert()
           |> with_block(block)
 
-        insert(:token_transfer, from_address: address, transaction: transaction)
+        insert(:token_transfer,
+          from_address: address,
+          transaction: transaction,
+          block: block,
+          block_number: block.number
+        )
       end
 
       options = %{
@@ -1135,7 +1195,12 @@ defmodule Explorer.EtherscanTest do
           |> insert()
           |> with_block(block)
 
-        insert(:token_transfer, from_address: address, transaction: transaction)
+        insert(:token_transfer,
+          from_address: address,
+          transaction: transaction,
+          block: block,
+          block_number: block.number
+        )
       end
 
       options = %{start_block: third_block.number}
@@ -1161,7 +1226,12 @@ defmodule Explorer.EtherscanTest do
           |> insert()
           |> with_block(block)
 
-        insert(:token_transfer, from_address: address, transaction: transaction)
+        insert(:token_transfer,
+          from_address: address,
+          transaction: transaction,
+          block: block,
+          block_number: block.number
+        )
       end
 
       options = %{end_block: second_block.number}
@@ -1190,7 +1260,14 @@ defmodule Explorer.EtherscanTest do
         |> with_block()
 
       insert(:token_transfer, from_address: address, transaction: transaction)
-      insert(:token_transfer, from_address: address, token_contract_address: contract_address, transaction: transaction)
+
+      insert(:token_transfer,
+        from_address: address,
+        token_contract_address: contract_address,
+        transaction: transaction,
+        block: transaction.block,
+        block_number: transaction.block_number
+      )
 
       [found_token_transfer] = Etherscan.list_token_transfers(address.hash, contract_address.hash)