+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Fixes
 - [#2871](https://github.com/poanetwork/blockscout/pull/2871) - do not remove logs
+- [#2872](https://github.com/poanetwork/blockscout/pull/2872) - do not remove token transfers
 - [#2800](https://github.com/poanetwork/blockscout/pull/2800) - return not found for not verified contract for token read_contract
 - [#2806](https://github.com/poanetwork/blockscout/pull/2806) - Fix blocks fetching on the main page
 - [#2803](https://github.com/poanetwork/blockscout/pull/2803) - Fix block validator custom tooltip